Add naive way of seeing if an infoWindow is open
This will break if we ever allow for more than one info window to be
open at any given time. Since we probably won't want to do that, it
is probably best to just make it so that there is only one info
window that exists and this will likely simplify things.