Refactor: Change to more basic script loading style
Revert to having the script tags at the bottom of the file. We have
removed the perhaps unnecessarily complex jsLoading.js script that
was inline in the header of index.html. This will need to be
tested to compare the speed different. Page Speed Insights likes
the old method more... is there an actual difference though?
What is a better solution?